D0c mute warn adjustments (#134)
* Updated Mute and Warn code

**Warn:**

- Provides minimal info in non-mod channels and provides user mention of user we warned in mod channel.

**Mute:** Resolves #124

- Keeps calling message and response in mod channel.
- Logs automatic un-mutes to mod channel.

* Fix build errors

* Further adjustments to Mute and Unmute

**Mute:**

- Users `this.client.user` to get user of self instead of constant

**Unmute:**

- Follows same rules as Mute; no longer deletes calling message

+++ src/commands/modtools/Mute.ts
@@ -120,10 +120,18 @@ export default class Mute extends Command<SweeperClient> {
 			this.client.mod.actions.mute(mutedUser, issuer, message.guild, muteTimeHUMN, note)
 				.then(result => {
 					this.client.mod.actions.setMuteDuration(mutedUser, message.guild, muteTimeMS);
-					message.delete();
 					this.logger.log('CMD Mute', `Muted user: '${mutedUser.user.tag}' in '${message.guild.name}'`);
 					try {
 						mutedUser.send(`You have been muted on **${message.guild.name}** for **${muteTimeHUMN}**.\n\n**A message from the mods:**\n\n"${note}"`);
+
+						// If message sent in the mod channel, then give full details, otherwise be vague
+						if (message.channel.id === Constants.modChannelId) {
+							message.channel.send(`Successfully muted <@${user.id}> for **${muteTimeHUMN}**.`);
+						} else {
+							message.channel.send(`That action was successful.`);
+							message.delete();
+						}
+
 					} catch (err) {
 						const modChannel: TextChannel = <TextChannel> message.guild.channels.get(Constants.modChannelId);
 						modChannel.send(`There was an error informing ${mutedUser.user.tag} of their mute. Their DMs may be disabled.\n\n**Error:**\n${err}`);
